% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_sql_connection.R
\name{delete_sql_connection}
\alias{delete_sql_connection}
\title{Delete a known SQL connection}
\usage{
delete_sql_connection(alias)
}
\arguments{
\item{alias}{string identifying the desired SQL connection.}
}
\description{
Function to delete a known SQL connection from your credential store.

If the given alias does not exist, an error is thrown.
}
\seealso{
Other connection management functions: 
\code{\link{create_sql_connection}()},
\code{\link{start_sql_connection}()}
}
\concept{connection management functions}
