% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_identifier.R
\name{sql_identifier}
\alias{sql_identifier}
\title{Create SQL identifiers}
\usage{
sql_identifier(name, parse = FALSE)
}
\arguments{
\item{name}{A string, \code{\link[DBI:SQL]{DBI::SQL()}} or \code{\link[DBI:Id]{DBI::Id()}} object to quote as an
identifier.}

\item{parse}{boolean; whether to attempt to parse \code{name} into constituent
parts. Only used if \code{name} is a character string.}
}
\value{
A \code{\link[DBI:SQL]{DBI::SQL()}} string properly identifying the desired object
}
\description{
A simple wrapper around \code{\link[DBI:dbQuoteIdentifier]{DBI::dbQuoteIdentifier()}}, see its documentation for
more information.
}
\details{
If \code{name} is an SQL string, it will be validated. If it fails the validation,
the function will throw an error.

If \code{name} is a character string and \code{parse = TRUE}, then the function will
attempt to parse it into constituent parts
(\verb{"db.schema.table" -> "db"."schema"."table"}). It will also be validated.
}
