% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_sql_connection.R
\name{start_sql_connection}
\alias{start_sql_connection}
\title{Start up the desired SQL connection}
\usage{
start_sql_connection(alias)
}
\arguments{
\item{alias}{string identifying the desired SQL connection.}
}
\value{
On success, a \code{pool::pool} object representing the desired
connection.

If the connection did not exist and the user cancels the setup, returns
\code{NULL} invisibly.
}
\description{
This function sets up the desired SQL connection if it has
been identified in the past. If not, then the connection is defined
interactively and set up accordingly.
}
\details{
This must be the first \code{simplysql} function called, since it sets up the
connection to be used everywhere else.

If the desired connection doesn't yet exist, it is created interactively,
with the user entering a valid \code{DBI::dbConnect()} call to generate the
desired connection.

This unusual interaction is meant to make debugging the connection easier.
The suggested workflow when making your first connection is as follows:

\itemize{
\item if you already have a \code{DBI::dbConnect()} call, simply copy-and-paste
it at the prompt and you're done.
\item if you've never connected to this database before, it'll be best to
validate your connection parameters with a simple \code{DBI::dbConnect()} call
first to make sure everything's right. This removes sources of
uncertainty in case anything goes wrong. Once you've managed to create a
valid connection, just copy that call and paste it at the prompt.
}

The \code{pool::pool} object responsible for handling the connection is
automatically registered to disconnect when the session is terminated.
}
\seealso{
Other connection management functions: 
\code{\link{create_sql_connection}()},
\code{\link{delete_sql_connection}()}
}
\concept{connection management functions}
