% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_queries.R
\name{exists_sql_table}
\alias{exists_sql_table}
\alias{list_sql_tables}
\alias{list_sql_table_fields}
\alias{list_sql_objects}
\title{Obtain structural data on the database}
\usage{
exists_sql_table(name, ...)

list_sql_tables(...)

list_sql_table_fields(name, ...)

list_sql_objects(prefix = NULL, ...)
}
\arguments{
\item{name}{The table name, passed on to \code{\link[DBI:dbQuoteIdentifier]{dbQuoteIdentifier()}}. Options are:
\itemize{
\item a character string with the unquoted DBMS table name,
e.g. \code{"table_name"},
\item a call to \code{\link[DBI:Id]{Id()}} with components to the fully qualified table name,
e.g. \code{Id(schema = "my_schema", table = "table_name")}
\item a call to \code{\link[DBI:SQL]{SQL()}} with the quoted and fully qualified table name
given verbatim, e.g. \code{SQL('"my_schema"."table_name"')}
}}

\item{...}{Other parameters passed on to methods.}

\item{prefix}{A fully qualified path in the database's namespace, or \code{NULL}.
This argument will be processed with \code{\link[DBI:dbUnquoteIdentifier]{dbUnquoteIdentifier()}}.
If given the method will return all objects accessible through this prefix.}
}
\description{
These functions simply wrap DBI functions to use the standard connection
defined in \code{\link[=start_sql_connection]{start_sql_connection()}}.
}
\details{
\itemize{
\item \code{exists_sql_table()} wraps \code{\link[DBI:dbExistsTable]{DBI::dbExistsTable()}}
\item \code{list_sql_tables()} wraps \code{\link[DBI:dbListTables]{DBI::dbListTables()}}
\item \code{list_sql_table_fields()} wraps \code{\link[DBI:dbListFields]{DBI::dbListFields()}}
\item \code{list_sql_objects()} wraps \code{\link[DBI:dbListObjects]{DBI::dbListObjects()}}
}

See the \code{DBI} documentation for further information on each function.

\code{name} will be simply interpreted by \code{\link[=sql_identifier]{sql_identifier()}}. If it is a character
string containing a fully qualified name (i.e. \code{my_schema.my_table}), call
\code{sql_identifier()} yourself and set \code{parse = TRUE}.
}
